requirements.txt
######################################################
#
# Python packages needed to run PyCosworth!
#
# Install with: 
# pip install -r requirements.txt   (if using Python 2.x)
# pip3 install -r requirements.txt   (if using Python 3.x)
#
######################################################

# Needed for any LCD/serial/terminal interface displays as well as ECU interface cable
pyserial

# Add the module to talk to an adafruit/matrix orbital LCD usb/serial interface
lcdbackpack

# GPIO modules as needed to read presses from buttons etc.
RPi.GPIO

# gpiozero has a nice method of determining how long buttons are pressed on the GPIO pins
# of the Raspberry Pi
gpiozero

# PIL is used to generate graphics on OLED modules or SDL windows on the desktop
Pillow

# Luma.oled supports many more OLED screens
luma.oled

# SDL2 is used to show graphics in the absence of any OLED screens
# We use pysdl2 as it is the most up to date SDL implementation for Python
pysdl2

# Numpy is used to do fast 2d pixel operations - needed to copy in-memory images, generated by PIL, to
# an SDL_Surface object
numpy
